Splits transforms, models, datasets from train into modules
